@@ -13,10 +13,10 @@
     "@dnd-kit/core": "^6.0.8",
     "@dnd-kit/modifiers": "^6.0.1",
     "@dnd-kit/sortable": "^8.0.0",
-    "@emotion/cache": "^11.10.5",
+    "@emotion/cache": "^11.11.0",
     "@emotion/is-prop-valid": "^1.1.1",
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
+    "@emotion/react": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
     "@mantine/core": "^6.0.13",
     "@mantine/dates": "^6.0.13",
     "@mantine/hooks": "^6.0.13",
@@ -25,6 +25,7 @@
     "@snowplow/browser-tracker": "^3.1.6",
     "@tanstack/react-virtual": "^3.1.2",
     "@tippyjs/react": "^4.2.6",
+    "@types/react-router": "3.0.27",
     "@types/redux-auth-wrapper": "^2.0.16",
     "@visx/axis": "1.8.0",
     "@visx/clip-path": "^2.1.0",
@@ -33,7 +34,7 @@
     "@visx/scale": "1.7.0",
     "@visx/shape": "2.12.2",
     "@visx/text": "1.7.0",
-    "ace-builds": "^1.4.13",
+    "ace-builds": "^1.4.14",
     "classnames": "^2.1.3",
     "color": "^4.2.3",
     "crc-32": "^1.2.2",
@@ -74,20 +75,20 @@
     "postcss-discard-comments": "^6.0.1",
     "prop-types": "^15.5.7",
     "re-reselect": "^4.0.1",
-    "react": "^17.0.2",
-    "react-ace": "^9.5.0",
+    "react": "^18.2.0",
+    "react-ace": "^10.1.0",
     "react-ansi-style": "^1.0.0",
     "react-beautiful-dnd": "^13.1.1",
     "react-color": "^2.14.1",
     "react-copy-to-clipboard": "^5.0.1",
     "react-dnd": "4",
     "react-dnd-html5-backend": "4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-draggable": "^4.4.5",
     "react-dropzone": "^14.2.3",
     "react-grid-layout": "^1.2.5",
     "react-innertext": "^1.1.5",
-    "react-is": "^17.0.2",
+    "react-is": "^18.2.0",
     "react-markdown": "^8.0.4",
     "react-motion": "0.5.2",
     "react-redux": "^8.0.5",
@@ -146,7 +147,7 @@
     "@swc/core": "^1.3.99",
     "@testing-library/cypress": "^9.0.0",
     "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/react-hooks": "^8.0.0",
     "@testing-library/user-event": "^14.5.2",
     "@types/babel__core": "^7.20.4",
@@ -186,17 +187,16 @@
     "@types/prettier": "1.19.1",
     "@types/prop-types": "^15.7.4",
     "@types/raf": "^3.4.0",
-    "@types/react": "^17.0.77",
+    "@types/react": "^18.2.42",
     "@types/react-beautiful-dnd": "^13.1.3",
     "@types/react-color": "^2.17.6",
     "@types/react-copy-to-clipboard": "^5.0.2",
-    "@types/react-dom": "^17.0.11",
+    "@types/react-dom": "^18.2.17",
     "@types/react-grid-layout": "^1.1.3",
-    "@types/react-is": "^17.0.3",
+    "@types/react-is": "^18.2.4",
     "@types/react-motion": "0.0.32",
     "@types/react-redux": "^7.1.24",
     "@types/react-resizable": "^3.0.4",
-    "@types/react-router": "3.0.28",
     "@types/react-router-redux": "^4.0.53",
     "@types/react-textarea-autosize": "^4.3.6",
     "@types/react-transition-group": "^4.4.5",
@@ -303,8 +303,8 @@
     "jsdom": "^22.1.0",
     "json5": "2.2.2",
     "nth-check": "2.0.1",
-    "@types/react": "^17.0.77",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.2.42",
+    "@types/react-dom": "^18.2.17",
     "set-value": "4.0.1",
     "source-map-js": "1.2.0",
     "unset-value": "2.0.1",
