@@ -17,7 +17,7 @@ export const PublicError = () => {
       >
         <NoDataError mb="1rem" />
         <div className={cx(CS.mt1, CS.h4, CS.smH3, CS.mdH2, CS.textBold)}>
-          {message}
+          {message.toString()}
         </div>
       </div>
     </SyncedEmbedFrame>
